Rename folder to be oci
